+++ cracker-client/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@apollo/client": "^3.1.3",
+    "@apollo/link-context": "^2.0.0-beta.3",
     "@auth0/auth0-react": "^1.0.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.0.3",
