+++ cracker-client/src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./+localization/i18n";
-import { client, ApolloProvider } from "./+utils/apolloSetup";
 import { BrowserRouter as Router } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 
@@ -9,21 +8,22 @@ import "./index.css";
 import { App } from "./App";
 import { theme } from "./+utils/theme";
 import { Auth0Provider } from "@auth0/auth0-react";
+import { GraphQLProvider } from "./+utils/GraphQLProvider";
 
 ReactDOM.render(
   <React.StrictMode>
     <Auth0Provider
       domain={process.env.REACT_APP_AUTH0_DOMAIN}
       clientId={process.env.REACT_APP_AUTH0_CLIENT_ID}
-      redirectUri={`${process.env.REACT_APP_AUTH0_ORIGIN}/callback`}
+      redirectUri={process.env.REACT_APP_AUTH0_ORIGIN}
     >
-      <ApolloProvider client={client}>
+      <GraphQLProvider>
         <Router>
           <ThemeProvider theme={theme}>
             <App />
           </ThemeProvider>
         </Router>
-      </ApolloProvider>
+      </GraphQLProvider>
     </Auth0Provider>
   </React.StrictMode>,
   document.getElementById("root")
